refactor: pricing model and bump nextjs
