+++ package.json
@@ -23,7 +23,7 @@
     "fumadocs-mdx": "^11.6.1",
     "fumadocs-ui": "^15.2.11",
     "lucide-react": "^0.471.1",
-    "next": "^15.3.0",
+    "next": "^15.3.3",
     "next-themes": "^0.4.4",
     "prism-themes": "^1.9.0",
     "prismjs": "^1.30.0",
@@ -49,5 +49,6 @@
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
-  }
+  },
+  "packageManager": "pnpm@10.12.1"
 }
